+++ packages/shared-browser/src/lib/index.ts
@@ -4,8 +4,16 @@ export { default as simulateKeyPress } from './keys/simulateKeyPress/simulateKey
 
 export { delay } from './time/delay/delay.ts';
 
+export { formatLocalDateTime } from './time/formatLocalDateTime/formatLocalDateTime.ts';
+
+export { generateZettelIdFromDate } from './time/generateZettelIdFromDate/generateZettelIdFromDate.ts';
+
 export { retryUntilTrue } from './loops/retryUntilTrue/retryUntilTrue.ts';
 
 export { mountSvelteComponent } from './svelte/mount/mountSvelteComponent.ts';
 
 export { mountSvelteComponents } from './svelte/mount/mountSvelteComponents.ts';
+
+export { cleanText } from './text/cleanText/cleanText.ts';
+
+export { decodeHtmlEntities } from './text/decodeHtmlEntities/decodeHtmlEntities.ts';