+++ packages/shared-browser/vitest.config.ts
@@ -4,6 +4,11 @@ export default defineConfig({
   test: {
     environment: 'node',
     globals: true,
+    exclude: [
+      '**/node_modules/**',
+      '**/dist/**',
+      '**/.svelte-kit/**', // 👈 ignore this
+    ],
     coverage: {
       provider: 'v8',
       reporter: ['text', 'html'],