+++ scripts/goodreads-obsidian-base-readlist/src/lib/Button.svelte
@@ -37,11 +37,5 @@
     &:hover {
       background-color: var(--color-background-wtr-hover);
     }
-
-    &:focus {
-      &:before {
-        content: '✅ ';
-      }
-    }
   }
 </style>