+++ scripts/goodreads-obsidian-base-readlist/src/lib/GrabDetails/GrabDetails.svelte
@@ -3,9 +3,13 @@
   import { getDatabase } from '../helpers/getDatabase';
   import { convertToMarkdown } from '../helpers/convertToMarkdown/convertToMarkdown';
   import Button from '../Button.svelte';
+  import ActionTooltip from '../ActionTooltip.svelte';
+
+  let tooltipTrigger = $state(0);
 
   async function onclick() {
     try {
+      tooltipTrigger++;
       const data = getDatabase();
 
       const markdown = convertToMarkdown(data);
@@ -19,6 +23,7 @@
 
 <div class="GrabDetails">
   <Button title="Grab details" {onclick}>Copy Details</Button>
+  <ActionTooltip description="Copied!" trigger={tooltipTrigger} />
 </div>
 
 <style lang="scss">